Restructure and improve user view page; Implement new project and experience views.
