+++ src/components/user/experience/row.tsx
@@ -1,20 +1,38 @@
 import Link from "next/link";
+import { useSession } from "next-auth/react";
 
 import { type UserExperience } from "@prisma/client";
 
 import { api } from "@utils/api";
 import SuccessBox from "@utils/success";
-
-import ReactMarkdown from "react-markdown";
+import { has_role } from "@utils/user/auth";
 
 const Row: React.FC<{
     experience: UserExperience
 }> = ({
     experience
 }) => {
-    const editUrl = "/user/profile/experiences?id=" + experience.id.toString();
-
+    const { data: session } = useSession();
     const deleteMut = api.user.deleteExperience.useMutation();
+
+    // Compile URLs.
+    const viewUrl = `/user/view/$${experience.userId}/experiences/${experience.id.toString()}`;
+    const editUrl = `/user/profile/experiences?id=${experience.id.toString()}`;
+
+    // See if we have permissions.
+    let canEdit = false;
+
+    if (session?.user) {
+        const userId = session.user.id;
+
+        // Is user owner?
+        if (userId == experience.userId)
+            canEdit = true;
+
+        // Is user admin or moderator?
+        if (has_role(session, "admin") || has_role(session, "moderator"))
+            canEdit = true;
+    }
 
     return (
         <>
@@ -29,25 +47,36 @@ const Row: React.FC<{
                         <h3>{experience.title}</h3>
                     </div>
                     <div className="experience-row-description">
-                        <ReactMarkdown
-                            className="markdown"
-                        >
-                            {experience.desc ?? ""}
-                        </ReactMarkdown>
+                        <p>{experience.desc ?? ""}</p>
                     </div>
+
                     <div className="experience-row-actions">
-                        <Link className="button button-primary" href={editUrl}>Edit</Link>
-                        <Link className="button button-danger" href="#" onClick={(e) => {
-                            e.preventDefault();
+                        <Link
+                            href={viewUrl}
+                            className="button"
+                        >View</Link>
+                        {canEdit && (
+                            <>
+                                <Link
+                                    href={editUrl}
+                                    className="button button-primary"
+                                >Edit</Link>
+                                <button
+                                    className="button button-danger"
+                                    onClick={(e) => {
+                                        e.preventDefault();
 
-                            const yes = confirm("Are you sure you want to delete this experience?");
+                                        const yes = confirm("Are you sure you want to delete this experience?");
 
-                            if (yes) {
-                                deleteMut.mutate({
-                                    id: experience.id
-                                });
-                            }
-                        }}>Delete</Link>
+                                        if (yes) {
+                                            deleteMut.mutate({
+                                                id: experience.id
+                                            });
+                                        }
+                                    }}
+                                >Delete</button>
+                            </>
+                        )}
                     </div>
                 </div>
             )}
