+++ src/pages/user/view/[user]/skills/index.tsx
@@ -1,0 +1,91 @@
+import { type GetServerSidePropsContext, type NextPage } from "next";
+
+import { type User } from "@prisma/client";
+
+import { prisma } from "@server/db";
+
+import Wrapper from "@components/wrapper";
+import Meta from "@components/meta";
+
+import SkillBrowser from "@components/user/skill/browser";
+import NotFound from "@components/errors/not_found";
+import UserView from "@components/user/view";
+
+import { type GlobalPropsType } from "@utils/global_props";
+
+const Page: NextPage<{
+    user?: User
+} & GlobalPropsType> = ({
+    user,
+
+    footerServices,
+    footerPartners
+}) => {
+    return (
+        <>
+            <Meta
+                title={`Skills - ${user?.name ?? "Not Found"} - Users - Deaconn`}
+            />
+            <Wrapper
+                footerServices={footerServices}
+                footerPartners={footerPartners}
+            >
+                <div className="content-item">
+                    {user ? (
+                        <UserView
+                            user={user}
+                            view="skills"
+                        >
+                            <div className="content-item">
+                                <h1>Skills</h1>
+                                <SkillBrowser
+                                    userId={user.id}
+                                    small={true}
+                                />
+                            </div>
+                        </UserView>
+                    ) : (
+                        <NotFound item="User" />
+                    )}
+                </div>
+            </Wrapper>
+        </>
+    );
+}
+
+export async function getServerSideProps(ctx: GetServerSidePropsContext) {
+    // Retrieve parameters.
+    const { params } = ctx;
+
+    const userId = params?.user?.toString();
+
+    let user: User | null = null;
+
+    if (userId) {
+        let lookup_url = true;
+
+        if (userId[0] == "$") {
+            lookup_url = false;
+        }
+
+        user = await prisma.user.findFirst({
+            where: {
+                ...(lookup_url && {
+                    url: userId
+                }),
+                ...(!lookup_url && {
+                    id: userId.substring(1)
+                })
+            }
+        });
+    }
+
+    return {
+        props: {
+            user: JSON.parse(JSON.stringify(user))
+        }
+    };
+}
+
+
+export default Page;
